refactor(deb): Split  DebDatasource in smaller peaces and prepare for multi-compression feature
