+++ lib/modules/datasource/deb/index.spec.ts
@@ -8,8 +8,12 @@ import { GlobalConfig } from '../../../config/global';
 import { hashStream, toSha256 } from '../../../util/hash';
 import type { GetPkgReleasesConfig } from '../types';
 import { cacheSubDir } from './common';
-import * as fileUtils from './file';
-import { getBaseSuiteUrl } from './url';
+import {
+  getBaseSuiteUrl,
+  getComponentUrl,
+  getPackageUrl,
+  getRegistryUrl,
+} from './url';
 import { DebDatasource } from '.';
 import { Fixtures } from '~test/fixtures';
 import * as httpMock from '~test/http-mock';
@@ -31,7 +35,6 @@ describe('modules/datasource/deb/index', () => {
   let extractedPackageFile: string;
 
   beforeEach(async () => {
-    vi.resetAllMocks();
     debDatasource = new DebDatasource();
     cacheDir = await dir({ unsafeCleanup: true });
     GlobalConfig.set({ cacheDir: cacheDir.path });
@@ -347,108 +350,6 @@ describe('modules/datasource/deb/index', () => {
       });
     });
   });
-
-  describe('downloadAndExtractPackage', () => {
-    it('should ignore error when fetching the InRelease content fails', async () => {
-      const packageArgs: [release: string, component: string, arch: string] = [
-        'stable',
-        'non-free',
-        'amd64',
-      ];
-
-      httpMock
-        .scope(debBaseUrl)
-        .get(getPackageUrl('', ...packageArgs))
-        .replyWithFile(200, fixturePackagesArchivePath2);
-      mockFetchInReleaseContent('wrong-hash', ...packageArgs, true);
-
-      await expect(
-        // TODO: method is private, so needs better testing
-        // eslint-disable-next-line @typescript-eslint/dot-notation
-        debDatasource['downloadAndExtractPackage'](
-          getComponentUrl(debBaseUrl, ...packageArgs),
-        ),
-      ).resolves.toEqual(
-        expect.objectContaining({
-          extractedFile: extractedPackageFile,
-          lastTimestamp: expect.anything(),
-        }),
-      );
-    });
-
-    it('should throw error when checksum validation fails', async () => {
-      httpMock
-        .scope(debBaseUrl)
-        .get(getPackageUrl('', 'bullseye', 'main', 'amd64'))
-        .replyWithFile(200, fixturePackagesArchivePath2);
-      mockFetchInReleaseContent('wrong-hash', 'bullseye', 'main', 'amd64');
-
-      await expect(
-        // TODO: method is private, so needs better testing
-        // eslint-disable-next-line @typescript-eslint/dot-notation
-        debDatasource['downloadAndExtractPackage'](
-          getComponentUrl(debBaseUrl, 'bullseye', 'main', 'amd64'),
-        ),
-      ).rejects.toThrow(`SHA256 checksum validation failed`);
-    });
-
-    it('should throw error for when extracting fails', async () => {
-      vi.spyOn(fileUtils, 'extract').mockRejectedValueOnce(new Error());
-
-      httpMock
-        .scope(debBaseUrl)
-        .get(getPackageUrl('', 'bullseye', 'main', 'amd64'))
-        .replyWithFile(200, fixturePackagesArchivePath2);
-      mockFetchInReleaseContent(
-        fixturePackagesArchiveHash2,
-        'bullseye',
-        'main',
-        'amd64',
-      );
-
-      await expect(
-        // TODO: method is private, so needs better testing
-        // eslint-disable-next-line @typescript-eslint/dot-notation
-        debDatasource['downloadAndExtractPackage'](
-          getComponentUrl(debBaseUrl, 'bullseye', 'main', 'amd64'),
-        ),
-      ).rejects.toThrow(`Missing metadata in extracted package index file!`);
-    });
-  });
-
-  describe('checkIfModified', () => {
-    it('should return true for different status code', async () => {
-      httpMock
-        .scope(debBaseUrl)
-        .head(getPackageUrl('', 'stable', 'non-free', 'amd64'))
-        .reply(200);
-
-      await expect(
-        // TODO: method is private, so needs better testing
-        // eslint-disable-next-line @typescript-eslint/dot-notation
-        debDatasource['checkIfModified'](
-          getPackageUrl(debBaseUrl, 'stable', 'non-free', 'amd64'),
-          new Date(),
-        ),
-      ).resolves.toBe(true);
-    });
-
-    it('should return true if request failed', async () => {
-      httpMock
-        .scope(debBaseUrl)
-        .head(getPackageUrl('', 'stable', 'non-free', 'amd64'))
-        .replyWithError('Unexpected Error');
-
-      await expect(
-        // TODO: method is private, so needs better testing
-        // eslint-disable-next-line @typescript-eslint/dot-notation
-        debDatasource['checkIfModified'](
-          getPackageUrl(debBaseUrl, 'stable', 'non-free', 'amd64'),
-          new Date(),
-        ),
-      ).resolves.toBe(true);
-    });
-  });
 });
 
 /**
@@ -500,7 +401,7 @@ function mockHttpCalls(
  * @param arch - The architecture (e.g., 'amd64').
  * @param error - Optional flag to simulate an error response (default is false).
  */
-function mockFetchInReleaseContent(
+export function mockFetchInReleaseContent(
   packageIndexHash: string,
   release: string,
   component: string,
@@ -529,66 +430,12 @@ function mockFetchInReleaseContent(
 }
 
 /**
- * Constructs a URL for accessing the component directory for a specific release and architecture.
- *
- * @param baseUrl - The base URL of the repository.
- * @param release - The release name or codename (e.g., 'buster', 'bullseye').
- * @param component - The component name (e.g., 'main', 'contrib', 'non-free').
- * @param arch - The architecture name (e.g., 'amd64', 'i386').
- * @returns The complete URL to the component directory.
- */
-function getComponentUrl(
-  baseUrl: string,
-  release: string,
-  component: string,
-  arch: string,
-): string {
-  return `${baseUrl}/debian/dists/${release}/${component}/binary-${arch}`;
-}
-
-/**
- * Constructs a URL for accessing the Packages.gz file for a specific component, release, and architecture.
- *
- * @param baseUrl - The base URL of the repository.
- * @param release - The release name or codename (e.g., 'buster', 'bullseye').
- * @param component - The component name (e.g., 'main', 'contrib', 'non-free').
- * @param arch - The architecture name (e.g., 'amd64', 'i386').
- * @returns The complete URL to the Packages.gz file.
- */
-function getPackageUrl(
-  baseUrl: string,
-  release: string,
-  component: string,
-  arch: string,
-) {
-  return `${getComponentUrl(baseUrl, release, component, arch)}/Packages.gz`;
-}
-
-/**
- * Constructs a URL used generating the component url with specific release, components, and architecture.
- *
- * @param baseUrl - The base URL of the repository.
- * @param release - The release name or codename (e.g., 'buster', 'bullseye').
- * @param components - An array of component names (e.g., ['main', 'contrib', 'non-free']).
- * @param arch - The architecture name (e.g., 'amd64', 'i386').
- * @returns The complete URL to the package registry.
- */
-function getRegistryUrl(
-  baseUrl: string,
-  release: string,
-  components: string[],
-  arch: string,
-) {
-  return `${baseUrl}/debian?suite=${release}&components=${components.join(',')}&binaryArch=${arch}`;
-}
-
-/**
  * Computes the SHA256 checksum of a specified file.
  *
  * @param filePath - path of the file
  * @returns resolves to the SHA256 checksum
  */
-function computeFileChecksum(filePath: string): Promise<string> {
+export function computeFileChecksum(filePath: string): Promise<string> {
   const stream = createReadStream(filePath);
   return hashStream(stream, 'sha256');
 }
