+++ lib/modules/datasource/deb/url.ts
@@ -1,4 +1,5 @@
 import { logger } from '../../../logger';
+import type { Http, HttpOptions } from '../../../util/http';
 import { joinUrlParts } from '../../../util/url';
 
 /**
@@ -92,4 +93,89 @@ function getReleaseParam(url: URL, optionalParams: string[]): string {
     }
   }
   throw new Error(`Missing one of suite query parameter`);
+}
+
+/**
+ * Checks if a packageUrl content has been modified since the specified timestamp.
+ *
+ * @param packageUrl - The URL to check.
+ * @param lastDownloadTimestamp - The timestamp of the last download.
+ * @returns True if the content has been modified, otherwise false.
+ * @see https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/If-Modified-Since
+ */
+export async function checkIfModified(
+  packageUrl: string,
+  lastDownloadTimestamp: Date,
+  http: Http,
+): Promise<boolean> {
+  const options: HttpOptions = {
+    headers: {
+      'If-Modified-Since': lastDownloadTimestamp.toUTCString(),
+    },
+  };
+
+  try {
+    const response = await http.head(packageUrl, options);
+    return response.statusCode !== 304;
+  } catch (error) {
+    logger.warn(
+      { packageUrl, lastDownloadTimestamp, errorMessage: error.message },
+      'Could not determine if package file is modified since last download',
+    );
+    return true; // Assume it needs to be downloaded if check fails
+  }
+}
+
+/**
+ * Constructs a URL for accessing the component directory for a specific release and architecture.
+ *
+ * @param baseUrl - The base URL of the repository.
+ * @param release - The release name or codename (e.g., 'buster', 'bullseye').
+ * @param component - The component name (e.g., 'main', 'contrib', 'non-free').
+ * @param arch - The architecture name (e.g., 'amd64', 'i386').
+ * @returns The complete URL to the component directory.
+ */
+export function getComponentUrl(
+  baseUrl: string,
+  release: string,
+  component: string,
+  arch: string,
+): string {
+  return `${baseUrl}/debian/dists/${release}/${component}/binary-${arch}`;
+}
+
+/**
+ * Constructs a URL for accessing the Packages.gz file for a specific component, release, and architecture.
+ *
+ * @param baseUrl - The base URL of the repository.
+ * @param release - The release name or codename (e.g., 'buster', 'bullseye').
+ * @param component - The component name (e.g., 'main', 'contrib', 'non-free').
+ * @param arch - The architecture name (e.g., 'amd64', 'i386').
+ * @returns The complete URL to the Packages.gz file.
+ */
+export function getPackageUrl(
+  baseUrl: string,
+  release: string,
+  component: string,
+  arch: string,
+): string {
+  return `${getComponentUrl(baseUrl, release, component, arch)}/Packages.gz`;
+}
+
+/**
+ * Constructs a URL used generating the component url with specific release, components, and architecture.
+ *
+ * @param baseUrl - The base URL of the repository.
+ * @param release - The release name or codename (e.g., 'buster', 'bullseye').
+ * @param components - An array of component names (e.g., ['main', 'contrib', 'non-free']).
+ * @param arch - The architecture name (e.g., 'amd64', 'i386').
+ * @returns The complete URL to the package registry.
+ */
+export function getRegistryUrl(
+  baseUrl: string,
+  release: string,
+  components: string[],
+  arch: string,
+): string {
+  return `${baseUrl}/debian?suite=${release}&components=${components.join(',')}&binaryArch=${arch}`;
 }
